Moved accordion to switch component as child
