+++ components/DataVerification/DataVerifItem.tsx
@@ -5,22 +5,6 @@ import { DataVerifSwitch, DataVerifAccordion } from ".";
 import { StepAttempt } from "../../types/CHAIN/StepAttempt";
 import { Accordion } from "react-native-paper/lib/typescript/src/components/List/List";
 
-const MOCK_PROMPT_OPTS = [
-	"No Prompt (Independent)",
-	"Shadow Prompt (approximately one inch)",
-	"Partial Physical Prompt (thumb and index finger)",
-	"Full Physical Prompt (hand-over-hand)",
-];
-
-const MOCK_BEHAV_OPTS = [
-	"Mild (did not interfere with task)",
-	"Moderate (interfered with task, but we were able to work through it)",
-	"Severe (we were not able to complete the task due to the severity of the behavior)",
-];
-
-const MOCK_PROMP_Q = "What prompt did you use to complete the step?";
-const MOCK_BEHAV_Q = "How severe was the challenging behavior?";
-
 type Props = {
 	stepAttempt: StepAttempt;
 };
@@ -33,23 +17,12 @@ const DataVerifItem: FC<Props> = ({ stepAttempt }) => {
 
 	const [behavSelected, setBehaveSelected] = useState(false);
 	const [completed, setCompleted] = useState(true);
-	const [behavAccordion, setBehavAccord] = useState(false);
-	const [promptAccordion, setPromptAccord] = useState(false);
 
 	const handleSwitchVal = (v: boolean, type: number) => {
 		if (type === 0) setCompleted(v);
 		if (type === 1) setBehaveSelected(v);
 		return v;
 	};
-
-	useEffect(() => {
-		setPromptAccord(!promptAccordion);
-		console.log(promptAccordion);
-	}, [completed]);
-
-	useEffect(() => {
-		setBehavAccord(!behavAccordion);
-	}, [behavSelected]);
 
 	/**
 	 * use context api, here:
@@ -67,21 +40,7 @@ const DataVerifItem: FC<Props> = ({ stepAttempt }) => {
 					<DataVerifSwitch
 						instruction={instruction}
 						type={QUESTION_TYPES.completion}
-						defaultValue={true}
 						id={stepId}
-						handleSwitchVal={handleSwitchVal}
-					/>
-				</View>
-				<View
-					style={[
-						styles.accordion,
-						{ display: behavSelected ? "flex" : "none" },
-					]}
-				>
-					<DataVerifAccordion
-						question={MOCK_PROMP_Q}
-						answerOptions={MOCK_PROMPT_OPTS}
-						active={promptAccordion}
 					/>
 				</View>
 			</View>
@@ -91,22 +50,7 @@ const DataVerifItem: FC<Props> = ({ stepAttempt }) => {
 					<DataVerifSwitch
 						instruction={instruction}
 						type={QUESTION_TYPES.challBehav}
-						defaultValue={false}
 						id={stepId}
-						handleSwitchVal={handleSwitchVal}
-					/>
-				</View>
-
-				<View
-					style={[
-						styles.accordion,
-						{ display: promptAccordion ? "flex" : "none" },
-					]}
-				>
-					<DataVerifAccordion
-						question={MOCK_BEHAV_Q}
-						answerOptions={MOCK_BEHAV_OPTS}
-						active={behavAccordion}
 					/>
 				</View>
 			</View>
